Add support for a device with just an rx and a tx software queue, add support for directly accessing tx/rx tcp bufs
For use cases where the actual details of the interface/phy like read/writes
etc.. has to be handled outside the smoltcp library, just provide a dummy
interface with just queues. The "headroom" parameter is useful in cases where
the user of smoltcp further wants to add encapsulations on top of the packet
(or gets the packet with encapsulations, which is the headroom in the rx queue)
and does not want to allocate more buffers for that and instead wants to use
the headroom space provided in the buffer

Also add support for direct access to the buffers used for tcp tx/rx. Nextensio
wants to work with large buffers at the same time ensuring that there is no buffer
thats queued up if it has no data in in, hence we provide APIs to swap out buffers